Готовый проект 11го спринта. Исправлены замечания после ревью
