+++ src/main/java/ru/yandex/practicum/filmorate/FilmorateApplication.java
@@ -8,7 +8,5 @@ public class FilmorateApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(FilmorateApplication.class, args);
-        String sql = "INSERT INTO films (name,description,release_date,duration) VALUES ('Clone','Serial','1990-10-01','150');";
-
     }
 }
